Indices fix

+++ test/IntegratedTests/Polarizability/test_PolarizabilityBath_integrated.cpp
@@ -30,9 +30,9 @@ TEST_CASE("Integrated PolarizabilityBath fulfills the omega_cut much smaller "
       fill::zeros); // double for result and asymptotic value
 
   // loop over indices
-  for (int i = 0; i < 3; i++) {
-    for (int j = 0; j < 3; j++) {
-      result(i, j) = pol.integrate_omega({(double)i, (double)j}, IM, omega_min,
+  for (size_t i = 0; i < 3; i++) {
+    for (size_t j = 0; j < 3; j++) {
+      result(i, j) = pol.integrate_omega({i, j}, IM, omega_min,
                                          omega_max, relerr, abserr);
     }
   }
@@ -88,15 +88,15 @@ TEST_CASE("Integrated PolarizabilityBath fulfills the omega_cut much larger "
   asymp_mat *= asymp;
 
   // loop over indices
-  for (int i = 0; i < 3; i++) {
-    for (int j = 0; j < 3; j++) {
-      result(i, j) = pol.integrate_omega({(double)i, (double)j}, IM, omega_min,
+  for (size_t i = 0; i < 3; i++) {
+    for (size_t j = 0; j < 3; j++) {
+      result(i, j) = pol.integrate_omega({i, j}, IM, omega_min,
                                          omega_a - 1e-3, relerr, abserr);
       result(i, j) +=
-          pol.integrate_omega({(double)i, (double)j}, IM, omega_a - 1e-3,
+          pol.integrate_omega({i, j}, IM, omega_a - 1e-3,
                               omega_a + 1e-3, relerr, abserr);
       result(i, j) +=
-          pol.integrate_omega({(double)i, (double)j}, IM, omega_a + 1e-3,
+          pol.integrate_omega({i, j}, IM, omega_a + 1e-3,
                               omega_max, relerr, abserr);
     }
   }
